Add new services for User, BoardTask and AuthService, and their mappings
This commit introduces services for performing CRUD operations on User and BoardTask entities, as well as services for authentication (AuthService). The new services are injected by defining them in the .NET Core DI container. AutoMapper is used for mapping domain models to DTOs (Data Transfer Objects) for appropriate data responses back to the client. The service methods use Mediator CQRS to dispatch commands and queries. The new BoardTask and User services are created to separate the application's business logic for these entities, making the code more readable and maintainable. Meanwhile, the AuthService is introduced to handle user authentication operations. Improving code modularity and encapsulation of functionalities helps to manage future updates and promotes code reusability.